Added icons
git-svn-id: http://find-object.googlecode.com/svn/trunk/find_object@50 620bd6b2-0a58-f614-fd9a-1bd335dccda9

+++ src/AddObjectDialog.cpp
@@ -175,7 +175,7 @@ void AddObjectDialog::setState(int state)
 			delete extractor;
 			if(selectedKeypoints.size() != (unsigned int)descriptors.rows)
 			{
-				printf("ERROR : keypoints=%lu != descriptors=%d\n", selectedKeypoints.size(), descriptors.rows);
+				printf("ERROR : keypoints=%d != descriptors=%d\n", (int)selectedKeypoints.size(), descriptors.rows);
 			}
 
 			CvRect roi = computeROI(selectedKeypoints);
